backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Michał Jagoda <michal-jagoda@o2.pl>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
djangorestframework = "^3.15.2"
redis = "^5.1.0"
celery = "^5.4.0"
openai = "^1.50.2"
opencv-python-headless = "^4.10.0.84"
fer = "^22.5.1"
deepface = "^0.0.93"
django-celery-beat = "^2.7.0"
pydub = "^0.25.1"
mtcnn = "^0.1.1"
mediapipe = "^0.10.15"
tf-keras = "^2.17.0"
textstat = "^0.7.4"
scikit-learn = "^1.5.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
